Renaming files to be consistent with new API
