+++ test-support/helpers/201-created/sync.js
@@ -1,8 +1,8 @@
 import wrapInExpectation from './utils/wrap-in-expectation';
 
-import rawExpectElement from   './raw/expect-element';
-import rawExpectNoElement from './raw/expect-no-element';
-import rawExpectComponent from './raw/expect-component';
+import rawExpectElement from   './raw/has-element';
+import rawExpectNoElement from './raw/has-no-element';
+import rawExpectComponent from './raw/has-component';
 
 var wrappedExpectElement   = wrapInExpectation(rawExpectElement);
 var wrappedExpectNoElement = wrapInExpectation(rawExpectNoElement);
