+++ tests/integration/has-component-hbs-integration-test.js
@@ -1,4 +1,4 @@
-import hasComponent from '../helpers/201-created/raw/expect-component';
+import hasComponent from '../helpers/201-created/raw/has-component';
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
